#544 Use translations for validation
