+++ client/src/dictionary/index.ts
